refactor : Section에 의존적인 메소드 Sections클래스 생성후 이동
