+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -5,7 +5,6 @@ import nextstep.subway.station.domain.Station;
 
 import javax.persistence.*;
 import java.util.*;
-import java.util.function.Predicate;
 
 @Entity
 public class Line extends BaseEntity {
@@ -16,8 +15,8 @@ public class Line extends BaseEntity {
     private String name;
     private String color;
 
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     public Line() {
     }
@@ -30,7 +29,7 @@ public class Line extends BaseEntity {
     public Line(String name, String color, Station upStation, Station downStation, int distance) {
         this.name = name;
         this.color = color;
-        sections.add(new Section(this, upStation, downStation, distance));
+        sections.addFirstSection(new Section(this, upStation, downStation, distance));
     }
 
     public void update(Line line) {
@@ -50,48 +49,8 @@ public class Line extends BaseEntity {
         return color;
     }
 
-    public List<Section> getSections() {
-        return sections;
-    }
-
-    public Station getUpStation() {
-        return sections.get(0).getUpStation();
-    }
-
-    public Station getDownStation() {
-        return sections.get(0).getDownStation();
-    }
-
-    public boolean isEmptySection() {
-        return sections.isEmpty();
-    }
-
-    public boolean isUnableRemoveStatus() {
-        return sections.size() <= 1;
-    }
-
-    public void removeSection(Section section) {
-        sections.remove(section);
-    }
-
-    public Optional<Section> getSection(Predicate<Section> sectionPredicate) {
-        return sections.stream()
-                .filter(sectionPredicate)
-                .findFirst();
-    }
-
-    public void updateUpStationIfSameUpStation(Section section) {
-        sections.stream()
-                .filter(it -> it.getUpStation() == section.getUpStation())
-                .findFirst()
-                .ifPresent(it -> it.updateUpStation(section.getDownStation(), section.getDistance()));
-    }
-
-    public void updateDownStationIfSameDownStation(Section section) {
-        sections.stream()
-                .filter(it -> it.getDownStation() == section.getDownStation())
-                .findFirst()
-                .ifPresent(it -> it.updateDownStation(section.getUpStation(), section.getDistance()));
+    private boolean isEmptySection() {
+        return sections.isEmptySections();
     }
 
     public List<Station> getStations() {
@@ -100,101 +59,24 @@ public class Line extends BaseEntity {
             return Arrays.asList();
         }
 
-        List<Station> stations = new ArrayList<>();
-        Station downStation = findUpStation();
-        stations.add(downStation);
-
-        while (downStation != null) {
-            Station finalDownStation = downStation;
-            Optional<Section> nextLineStation = getSection(section -> section.isSameUpStation(finalDownStation));
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-            downStation = nextLineStation.get().getDownStation();
-            stations.add(downStation);
-        }
-
-        return stations;
-    }
-
-    private Station findUpStation() {
-        Collections.sort(sections);
-
-        Station downStation = getUpStation();
-        while (downStation != null) {
-            Station finalDownStation = downStation;
-
-            Optional<Section> nextLineStation = getSection(section -> section.isSameDownStation(finalDownStation));
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-            downStation = nextLineStation.get().getUpStation();
-        }
-
-        return downStation;
+        return sections.getStations();
     }
 
     public void addSection(Section section) {
-        List<Station> stations = getStations();
         section.setLine(this);
-
-        validationSection(section, stations);
-
-        boolean isUpStationExisted = isExistStation(stations, section.getUpStation());
-        boolean isDownStationExisted = isExistStation(stations, section.getDownStation());
-
-        if (isFirstRegistSection(section, stations)) {
+        if (sections.isEmptySections()) {
+            sections.addFirstSection(section);
             return;
         }
 
-        if (isUpStationExisted) {
-            updateUpStationIfSameUpStation(section);
-        } else if (isDownStationExisted) {
-            updateDownStationIfSameDownStation(section);
-        }
-        sections.add(section);
-    }
-
-    private boolean isFirstRegistSection(Section section, List<Station> stations) {
-        if (stations.isEmpty()) {
-            sections.add(section);
-            return true;
-        }
-        return false;
-    }
-
-    private void validationSection(Section section, List<Station> stations) {
-        boolean isUpStationExisted = isExistStation(stations, section.getUpStation());
-        boolean isDownStationExisted = isExistStation(stations, section.getDownStation());
-
-        if (!(isUpStationExisted ^ isDownStationExisted)) {
-            throw new RuntimeException("이미 등록된 구간 또는 등록할수 없는 구간입니다.");
-        }
-
-        if (!stations.isEmpty() && stations.stream().noneMatch(it -> it == section.getUpStation()) &&
-                stations.stream().noneMatch(it -> it == section.getDownStation())) {
-            throw new RuntimeException("등록할 수 없는 구간 입니다.");
-        }
-    }
-
-    private boolean isExistStation(List<Station> stations, Station station) {
-        return stations.stream().anyMatch(it -> it == station);
+        sections.addSection(section);
     }
 
     public void removeLineStation(Station station) {
-        if (isUnableRemoveStatus()) {
-            throw new RuntimeException();
-        }
-        Optional<Section> upLineStation = getSection(section -> section.isSameUpStation(station));
-        Optional<Section> downLineStation = getSection(section -> section.isSameDownStation(station));
+        Section newSection = sections.removeLineStation(station);
 
-        if (upLineStation.isPresent() && downLineStation.isPresent()) {
-            Station newUpStation = downLineStation.get().getUpStation();
-            Station newDownStation = upLineStation.get().getDownStation();
-            int newDistance = upLineStation.get().getDistance() + downLineStation.get().getDistance();
-            getSections().add(new Section(this, newUpStation, newDownStation, newDistance));
+        if(newSection != null) {
+            newSection.setLine(this);
         }
-        upLineStation.ifPresent(it -> removeSection(it));
-        downLineStation.ifPresent(it -> removeSection(it));
     }
 }
